+++ CubeQuest/Assets/scripts/playerscript.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class playerscript : MonoBehaviour {
     //vars
     public int lives; //extras chances
-    public static int HP = 50; //character health
-    public int MAXHP = 50;
+    public static float HP = 50; //character health
+    public float MAXHP = 50;
     public int currentitem; //enum? each weapon type can have a number associated with it.
     public float movespeed = 1000f; //speed multiplier applied to unit vector for motion
     public Vector3 currentpos;
